Make mail sending options configurable
The username, password, to, from, cc, and SMTP server URL are all now
settable via the config file (defaults to /etc/sprinklerd.conf). The
variables in the config file are:

emailUsername
emailPassword
emailURL
emailTo
emailFrom
emailCC

If any of the first 4 are missing, it will not send email, as it
considers those to be critical. I left the From off the list because I
wasn't sure if that also fit into the critical category or not; might
make sense to add it. The CC is optional; if it's left empty, it will
not include it in the sent email.

The URL should be of the form:

smtps://my.domain.name:465

SMTPS is not required, but the protocol should match whatever protocol
is actually being served on the port provided.