Modify which type of Thrift server is used and add logging
We were using the TSimpleServer and has been changed to the TThreadedServer. I
noticed I was having some problems with the server becoming unresponsive to
additional client connections. It seemed that a client connection was hanging
around and the server was waiting for it to nicely disconnect, which it did not.

I'm not quite sure why it was hanging around; there wasn't anything to indicate
that it should be in the webserver code. At any rate, I switched to the threaded
server to try to bypass the problem. It seems to have mostly worked. Additionally,
I added some logging in an effort to find out where it was getting stuck. I have
since commented the logging out because it looks like on the version of raspbian I
have, that syslog can deadlock (livelock? not sure). I put the sprinklerdaemon in
a debugger and saw that it had 15 threads all waiting to log a message to syslog.
After removing the syslog calls things seem to be more stable and I haven't seen
anymore locking.