Add timeouts and logging to web service
In an effort to better show the user what's going on, I've added a
timeout to the client side of the connection, and included some more
"logging" output (it's really just system printlns). I also added a
small div on the login page to try to indicate what the problem could be
for the user. Right now, if it's not working the user is pretty well in
the dark about what could be happening.